Velocity events work, I'm just stupid - fixed similar issue with Bungee

+++ common/src/main/java/ca/sperrer/p0t4t0sandwich/playtimeutils/common/storage/MongoDBDataSource.java
@@ -111,7 +111,7 @@ public class MongoDBDataSource implements DataSource {
             }
 
             // Get last streak
-            long last_streak = Long.valueOf(player_data.getLong("last_streak"));
+            long last_streak = Long.valueOf(player_data.getInteger("last_streak"));
             long timeval = unixTime - last_streak - (unixTime % 86400);
             streak = player_data.getInteger("streak");
